Include PoBA artifact
Version the Truffle artifact of the PoBA contract so that the deployed
Rinkeby version is available to be used, and we don't have to deploy it
again to try things on a testnet.

This contract was deployed by the same account that is used when running
ganache, so the private key for the server is the same one.